backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-dotenv = "^1.0.1"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
nvidia-ml-py3 = "^7.352.0"
boto3 = "^1.35.49"
pinecone-client = "^5.0.1"
llama-index = "0.11.20"
llama-index-embeddings-nvidia = "^0.2.4"
llama-index-llms-nvidia = "^0.2.6"
llama-index-vector-stores-pinecone = "^0.2.1"
pypdf2 = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
